Refactored MainWindow
* Moved MainWindow into the VGraph.src.ui namespace to make things  a little less spread out all over the place.
