+++ src/ui/MainWindow.xaml
@@ -5,11 +5,11 @@
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:skia="clr-namespace:SkiaSharp.Views.WPF;assembly=SkiaSharp.Views.WPF"
         xmlns:ui="clr-namespace:VGraph.src.ui"
-        x:Class="VGraph.MainWindow" 
-    mc:Ignorable="d"
-        Title="VGraph" Height="450" Width="800" KeyDown="VGraphWindow_OnKeyDown">
+        x:Class="VGraph.src.ui.MainWindow" 
+        mc:Ignorable="d"
+        Title="VGraph" Height="450" Width="800">
     <Window.CommandBindings>
-        <CommandBinding Command="ui:MenuCommands.NewGridCmd" Executed="ui:MenuBarControl.NewCommandBinding_Executed"/>
+        <CommandBinding Command="ui:MenuCommands.NewGridCmd" Executed="NewCommandBinding_Executed"/>
         <CommandBinding Command="ui:MenuCommands.OpenGridCmd" Executed="OpenCommandBinding_Executed"/>
         <CommandBinding Command="ui:MenuCommands.SaveGridCmd" Executed="SaveCommandBinding_Executed"/>
         <CommandBinding Command="ui:MenuCommands.ExportGridCmd" Executed="ExportCommandBinding_Executed"/>
